To-Do Angular CRUD Application
I build a To-Do Angular CRUD Application, learning to integrate the front-end with a back-end database. I implemented persistent storage applications as well as relational database applications:  a Create operation that allowed me to add a new record to a table;  a Read operation that allowed me to read record details; Update operation that allowed me to modify an existing record; a Delete operation that allowed me to delete any specific record based on the primary key.